requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Fiona==1.9.4.post1
Flask==2.3.2
Flask-Cors==3.0.10
GeoAlchemy2==0.13.3
geopandas==0.13.2
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
packaging==23.1
pandas==2.0.2
psycopg2==2.9.6
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.5
zipp==3.15.0
